# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CarePulse HTTP API
  version: 0.3.0
  description: |
    The HTTP surface of the CarePulse app: route handlers the deployment and
    external services call. This spec covers the endpoint routes only.

    Every *mutation* in the app is a Next.js server action (`"use server"`),
    not a route in this spec: those are validated, authorised, rate-limited and
    audited by the action modules (see `docs/adr/ADR-007`). Route handlers are
    the thin HTTP layers over the same modules — the SMS drain cron and the
    document stream.

    Session auth is cookie-based (`hcs_admin_session` / `hcs_patient_session`,
    see `docs/adr/ADR-002`). The cron endpoint does not use sessions; it
    authenticates with the `x-cron-secret` header and fails closed when the
    deployment has not configured `CRON_SECRET`.

    Interactive rendering: paste this file into https://editor.swagger.io or
    open it with any OpenAPI viewer (Redoc, Swagger UI).
servers:
  - url: http://localhost:3100
    description: Local development (matches the e2e web server port)
paths:
  /api/cron/jobs:
    get:
      operationId: drainJobQueue
      summary: Drain the persistent job queue
      description: >
        Seeds today's reminder-dispatch job (a deterministic day-scoped id, so
        repeated calls cannot stack dispatches) and then runs up to
        `JOB_QUEUE_DRAIN_LIMIT` due jobs: `invoice.generate` builds a PDF, uploads
        it to the document bucket, audits the creation, and notifies the patient;
        `reminder.dispatch` scans for scheduled appointments in the coming 24h and
        enqueues a de-duplicated SMS per patient.
      security:
        - cronSecret: []
      responses:
        "200":
          description: Queue drained
          content:
            application/json:
              schema:
                type: object
                required: [ok, processed]
                properties:
                  ok:
                    type: boolean
                    const: true
                  processed:
                    type: array
                    items:
                      type: object
                      required: [$id, type, outcome, attempts]
                      properties:
                        $id: { type: string }
                        type: { type: string }
                        outcome:
                          type: string
                          enum: [done, retry, failed]
                        attempts: { type: integer }
                        result: { type: string }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/CronDisabled"
...
components:
  securitySchemes:
    cronSecret:
      description: |
        Authenticates the scheduled drain endpoints. The caller sends the
        deployment's `CRON_SECRET` in the `x-cron-secret` header; a missing or
        mismatched secret is rejected with 401/503, so an unconfigured
        deployment never exposes a public drain.
      type: apiKey
      in: header
      name: x-cron-secret
    sessionCookie:
      description: |
        The browser-authenticated HTTP API. Auth comes from the signed session
        cookies `hcs_admin_session` / `hcs_patient_session`; route handlers call
        `getSession()` and enforce role + ownership server-side.
      type: apiKey
      in: cookie
      name: hcs_admin_session
...
  responses:
    Unauthorized:
      description: Missing or non-matching `x-cron-secret`.
      content:
        application/json:
          schema:
            type: object
            required: [ok, error]
            properties:
              ok:
                type: boolean
                const: false
              error:
                type: string
    CronDisabled:
      description: "`CRON_SECRET` is not configured; the drain is disabled."
      content:
        application/json:
          schema:
            type: object
            required: [ok, error]
            properties:
              ok:
                type: boolean
                const: false
              error:
                type: string
    InternalError:
      description: The drain itself failed.
      content:
        application/json:
          schema:
            type: object
            required: [ok, error]
            properties:
              ok:
                type: boolean
                const: false
              error:
                type: string
...